workspace/buff_detector.py
import cv2
import numpy as np
import os

class BuffDetector:
    def __init__(self, coordinates):
        self.coordinates = coordinates
        self.templates = self.load_templates()

    def load_templates(self):
        templates = {}
        for filename in os.listdir('buffs'):
            if filename.endswith('.png'):
                templates[filename] = cv2.imread('buffs/' + filename, 0)
        return templates

    def detect(self):
        # Capture screen
        screen = self.capture_screen()

        # Detect buffs
        buffs = []
        for name, template in self.templates.items():
            result = cv2.matchTemplate(screen, template, cv2.TM_CCOEFF_NORMED)
            _, _, _, max_loc = cv2.minMaxLoc(result)
            buffs.append((name, max_loc))

        return buffs

    def capture_screen(self):
        # Placeholder for screen capture functionality
        return np.zeros((1080, 1920), dtype=np.uint8)
