workspace/requirements.txt
numpy
opencv-python
Pillow
pytesseract
